Update package.json and add lib.ts
