+++ .npmrc
@@ -1,1 +1,3 @@
 save-prefix=""
+# verdaccio registry
+registry="http://localhost:4873"
